fuzz: Update to latest dev, add IP6 support
