+++ test/linux/fuzz/continous_fuzz.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 set -e
-export CC=clang-7.0
-export CXX=clang++-7.0
+export CC=clang-9
+export CXX=clang++-9
 RUN=OFF $INCLUDEOS_PREFIX/bin/lxp-run
 BINARY=build/"`cat build/binary.txt`"
 # use -help=1 to see parameters to libfuzzer
-$BINARY -max_len=120
+LD_LIBRARY_PATH=$HOME/llvm/install/lib $BINARY -max_len=120
