+++ test/linux/fuzz/fuzzy_packet.hpp
@@ -18,4 +18,11 @@ namespace fuzzy
   extern uint8_t*
   add_tcp4_layer(uint8_t* data, FuzzyIterator& fuzzer,
                 const uint16_t dport);
+
+  extern uint8_t*
+  add_ip6_layer(uint8_t* data, FuzzyIterator& fuzzer,
+                const net::ip6::Addr src_addr,
+                const net::ip6::Addr dst_addr,
+                const uint8_t protocol);
+
 }
